Split MetaIterConfig to its own file.  Fix stale cache of config object in Effects
